+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 # Change Log
 
-## [master] - 2020-03-22
+## [master] - 2020-04-01
 
 ### Changed
 
 - Use `CPU_ONLY` compile when `torch` fails to detect a GPU (Issue #105)
 - Fix `get_kernel_map` for `CPU_ONLY` (Issue #107)
 - Update `get_union_map` doc (Issue #108)
+- Abstract getattr minkowski backend functions
 
 
 ## [0.4.2] - 2020-03-13